Cleaning up IO code
