+++ synaptor/io/__init__.py
@@ -1,4 +1,6 @@
-from .dispatch import *
-from .cloudvolume import *
+from . import backends
+from .backends.cloudvolume import *
+from .backends.utils import *
 
-from . import utils
+from .base import *
+from .utils import *
