+++ synaptor/io/backends/__init__.py
@@ -1,0 +1,5 @@
+from . import aws
+from . import gcloud
+from . import cloudvolume
+from . import local
+from . import utils
